requirements.txt
pyyaml>=6.0
requests>=2.31
openai>=1.0
pdfplumber>=0.10
pypdf>=3.0
python-docx>=1.0
google-api-python-client>=2.100
google-auth>=2.23
